+++ packages/ethereum-contracts/dev-scripts/run-deploy-contracts-and-token.js
@@ -18,6 +18,8 @@ deployContractsAndToken()
             nativeAssetSuperTokenAddress:
                 tokenDeploymentOutput.nativeAssetSuperTokenData
                     .nativeAssetSuperTokenAddress,
+            constantOutflowNFTAddress: frameworkAddresses.constantOutflowNFT,
+            constantInflowNFTAddress: frameworkAddresses.constantInflowNFT,
         };
 
         // create json output