+++ packages/sdk-core/src/ERC20Token.ts
@@ -2,7 +2,13 @@ import { ethers } from "ethers";
 
 import Operation from "./Operation";
 import { SFError } from "./SFError";
-import { IBaseSuperTokenParams, ITransferFromParams } from "./interfaces";
+import {
+    ERC20AllowanceParams,
+    ERC20BalanceOfParams,
+    IBaseSuperTokenParams,
+    ITransferFromParams,
+    ProviderOrSigner,
+} from "./interfaces";
 import { IERC20Metadata, IERC20Metadata__factory } from "./typechain-types";
 import { normalizeAddress } from "./utils";
 
@@ -32,11 +38,7 @@ export default class ERC20Token {
         owner,
         spender,
         providerOrSigner,
-    }: {
-        owner: string;
-        spender: string;
-        providerOrSigner: ethers.providers.Provider | ethers.Signer;
-    }): Promise<string> => {
+    }: ERC20AllowanceParams): Promise<string> => {
         const normalizedOwner = normalizeAddress(owner);
         const normalizedSpender = normalizeAddress(spender);
         try {
@@ -62,10 +64,7 @@ export default class ERC20Token {
     balanceOf = async ({
         account,
         providerOrSigner,
-    }: {
-        account: string;
-        providerOrSigner: ethers.providers.Provider | ethers.Signer;
-    }): Promise<string> => {
+    }: ERC20BalanceOfParams): Promise<string> => {
         try {
             const normalizedAccount = normalizeAddress(account);
             const balanceOf = await this.contract
@@ -89,7 +88,7 @@ export default class ERC20Token {
     name = async ({
         providerOrSigner,
     }: {
-        providerOrSigner: ethers.providers.Provider | ethers.Signer;
+        providerOrSigner: ProviderOrSigner;
     }): Promise<string> => {
         try {
             const name = await this.contract.connect(providerOrSigner).name();
@@ -111,7 +110,7 @@ export default class ERC20Token {
     symbol = async ({
         providerOrSigner,
     }: {
-        providerOrSigner: ethers.providers.Provider | ethers.Signer;
+        providerOrSigner: ProviderOrSigner;
     }): Promise<string> => {
         try {
             const symbol = await this.contract
@@ -135,7 +134,7 @@ export default class ERC20Token {
     totalSupply = async ({
         providerOrSigner,
     }: {
-        providerOrSigner: ethers.providers.Provider | ethers.Signer;
+        providerOrSigner: ProviderOrSigner;
     }): Promise<string> => {
         try {
             const totalSupply = await this.contract