+++ packages/sdk-core/src/interfaces.ts
@@ -13,6 +13,8 @@ import {
 // Maybe moving these into categorical files
 // makes more sense than stuffing them all here
 
+export type ProviderOrSigner = ethers.providers.Provider | ethers.Signer;
+
 // read request interfaces
 export interface IAccountTokenSnapshotFilter {
     readonly account?: string;
@@ -88,16 +90,16 @@ export interface ISuperTokenGetSubscriptionParams {
     readonly indexId: string;
     readonly publisher: string;
     readonly subscriber: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 export interface ISuperTokenGetIndexParams {
     readonly indexId: string;
     readonly publisher: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 export interface ISuperTokenPublisherParams extends ISuperTokenBaseIDAParams {
     readonly publisher: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 export interface ISuperTokenPubSubParams extends EthersParams {
     readonly indexId: string;
@@ -178,7 +180,7 @@ export interface IFullControlParams
 }
 
 export interface IRealtimeBalanceOfParams {
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
     readonly account: string;
     readonly timestamp?: number;
 }
@@ -203,53 +205,53 @@ export interface ERC777SendParams extends EthersParams {
 export interface ISuperTokenGetFlowParams {
     readonly sender: string;
     readonly receiver: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface ISuperTokenGetFlowInfoParams {
     readonly account: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IGetFlowParams {
     readonly superToken: string;
     readonly sender: string;
     readonly receiver: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IGetAccountFlowInfoParams {
     readonly superToken: string;
     readonly account: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IGetFlowOperatorDataParams {
     readonly superToken: string;
     readonly sender: string;
     readonly flowOperator: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IGetFlowOperatorDataByIDParams {
     readonly superToken: string;
     readonly flowOperatorId: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IGetGovernanceParametersParams {
-    providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    providerOrSigner: ProviderOrSigner;
     token?: string;
 }
 export interface ISuperTokenFlowOperatorDataParams {
     readonly sender: string;
     readonly flowOperator: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface ISuperTokenFlowOperatorDataByIDParams {
     readonly flowOperatorId: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IBaseIDAParams {
@@ -271,11 +273,11 @@ export interface IBaseSubscriptionParams extends IBaseIDAParams {
 export interface IGetSubscriptionParams extends IBaseIDAParams {
     readonly publisher: string;
     readonly subscriber: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 export interface IGetIndexParams extends IBaseIDAParams {
     readonly publisher: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface IDistributeParams extends EthersParams {
@@ -516,15 +518,15 @@ export interface IWeb3GovernanceParams {
     readonly minimumDeposit: string;
 }
 
-export interface ERC20BalanceOfParams extends EthersParams {
+export interface ERC20BalanceOfParams {
     readonly account: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
-export interface ERC20AllowanceParams extends EthersParams {
+export interface ERC20AllowanceParams {
     readonly owner: string;
     readonly spender: string;
-    readonly providerOrSigner: ethers.providers.Provider | ethers.Signer;
+    readonly providerOrSigner: ProviderOrSigner;
 }
 
 export interface ERC20ApproveParams extends EthersParams {
@@ -544,6 +546,52 @@ export interface ERC20TransferFromParams extends ERC20TransferParams {
     readonly signer: ethers.Signer;
 }
 
+// ERC721
+
+export interface NFTFlowData {
+    readonly flowSender: string;
+    readonly flowStartDate: Date;
+    readonly flowReceiver: string;
+}
+
+export interface ERC721TransferFromParams extends EthersParams {
+    readonly from: string;
+    readonly to: string;
+    readonly tokenId: string;
+}
+
+export interface ERC721SafeTransferFromParams extends ERC721TransferFromParams {
+    readonly data: string;
+}
+
+export interface ERC721ApproveParams extends EthersParams {
+    readonly approved: string;
+    readonly tokenId: string;
+}
+
+export interface ERC721SetApprovalForAllParams extends EthersParams {
+    readonly operator: string;
+    readonly approved: boolean;
+}
+
+export interface ERC721BalanceOfParams {
+    readonly owner: string;
+    readonly providerOrSigner: ProviderOrSigner;
+}
+
+export interface ERC721TokenIdQueryParams {
+    readonly tokenId: string;
+    readonly providerOrSigner: ProviderOrSigner;
+}
+export interface ERC721IsApprovedForAllParams {
+    readonly owner: string;
+    readonly operator: string;
+    readonly providerOrSigner: ProviderOrSigner;
+}
+
+export type ERC721OwnerOfParams = ERC721TokenIdQueryParams;
+export type ERC721GetApprovedParams = ERC721TokenIdQueryParams;
+export type ERC721TokenURIParams = ERC721TokenIdQueryParams;
 export interface ERC20IncreaseAllowanceParams extends EthersParams {
     readonly spender: string;
     readonly amount: string;