+++ packages/subgraph/scripts/getAbi.js
@@ -4,6 +4,7 @@ const path = require("path");
 const contracts = [
     "ERC20",
     "IConstantFlowAgreementV1",
+    "IFlowNFTBase",
     "ISuperTokenFactory",
     "ISuperToken",
     "ISuperfluid",